DBから住所取得し、jsに渡す

+++ src/main/java/jp/co/rakus/jazz/domain/Bar.java
@@ -9,8 +9,8 @@ package jp.co.rakus.jazz.domain;
 //`tel` VARCHAR(45) NULL,
 //`region_id` INT NOT NULL,
 //`prefecture_id` INT NOT NULL,
-//`latitude` INT,
-//`longitude` INT,
+//`latitude` double(8,6) DEFAULT NULL,
+//`longitude` double(8,6) DEFAULT NULL,
 //PRIMARY KEY (`id`),
 //UNIQUE INDEX `id_UNIQUE` (`id` ASC),
 //UNIQUE INDEX `tel_UNIQUE` (`tel` ASC));
@@ -32,80 +32,64 @@ public class Bar {
 	/** 都道府県 */
 	private Integer prefectureId;
 	/** 緯度 */
-	private Integer latitude;
+	private double latitude;
 	/** 経度 */
-	private Integer longitude;
-
+	private double longitude;
+	
 	public Integer getId() {
 		return id;
 	}
-
 	public void setId(Integer id) {
 		this.id = id;
 	}
-
 	public String getNameJpa() {
 		return nameJpa;
 	}
-
 	public void setNameJpa(String nameJpa) {
 		this.nameJpa = nameJpa;
 	}
-
 	public String getNameEng() {
 		return nameEng;
 	}
-
 	public void setNameEng(String nameEng) {
 		this.nameEng = nameEng;
 	}
-
 	public String getAddress() {
 		return address;
 	}
-
 	public void setAddress(String address) {
 		this.address = address;
 	}
-
 	public String getTel() {
 		return tel;
 	}
-
 	public void setTel(String tel) {
 		this.tel = tel;
 	}
-
 	public Integer getRegionId() {
 		return regionId;
 	}
-
 	public void setRegionId(Integer regionId) {
 		this.regionId = regionId;
 	}
-
 	public Integer getPrefectureId() {
 		return prefectureId;
 	}
-
 	public void setPrefectureId(Integer prefectureId) {
 		this.prefectureId = prefectureId;
 	}
-
-	public Integer getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-
-	public void setLatitude(Integer latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-
-	public Integer getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-
-	public void setLongitude(Integer longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
+	
 
 }
